refactor(runners): extract gemini turn runner to runners/gemini.rs (Phase 2c)
run_gemini_turn plus the Google credential resolution helpers
(GeminiCredentials, oauth/opencode-auth/env lookups) move verbatim.
Session-summary and debug-file helpers stay shared in mission_runner.
No behavior change.
